tests/wpt/tests/css/css-overflow/root-scroll-button-activation.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Overflow: ::scroll-button on root element activation</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow-5/#scroll-buttons">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  body {
    margin: 0;
  }

  :root::scroll-button(block-end) {
    content: "down";
    position: absolute;
    top: 0;
  }

  div {
    width: 600px;
    height: 300px;
    margin-bottom: 20px;
    background: green;
  }
</style>
<div></div>
<div></div>
<div></div>
<script>
  promise_test(async t => {
    await new test_driver.Actions()
    .pointerMove(10, 10)
    .pointerDown()
    .pointerUp()
    .pointerDown()
    .pointerUp()
    .send();
    assert_equals(window.scrollY, 360);
  });
</script>
